Add some rotation-related functions + test.
